requirements.txt
altair==4.2.2
attr==0.3.2
attrs==22.2.0
brotli==1.0.9
ConfigParser==5.3.0
cryptography==40.0.2
Cython==0.29.34
dl==0.1.0
docutils==0.19
HTMLParser==0.0.2
importlib_metadata==6.0.0
ipython==8.13.1
ipywidgets==8.0.6
Jinja2==3.1.2
jnius==1.1.0
keyring==23.13.1
lockfile==0.12.2
lxml==4.9.2
numpy==1.24.2
ordereddict==1.1
pandas==1.5.3
Pillow==9.5.0
protobuf==4.22.3
pyOpenSSL==23.1.1
railroad==0.5.0
Sphinx==7.0.0
streamlit==1.21.0
toml==0.10.2
tornado==6.2
xmlrpclib==1.0.1
zipp==3.15.0
